+++ .husky/pre-commit
@@ -1,4 +1,3 @@
-cd rs-react-app
 npm run lint
 npm test -- --run
