+++ src/pages/Home.tsx
@@ -4,9 +4,10 @@ import { useSearchParams } from 'react-router-dom';
 import { DetailsView } from '../svalka/DetailsView.tsx';
 import { FallbackUI } from '../svalka/FallbackUI.tsx';
 import { SearchResults } from '../svalka/SearchResults.tsx';
-import { ErrorBoundary } from '../../features/error/ErrorBoundary.tsx';
+
 
 import type { CharacterData } from '../utils/hooks/useCharacterData.ts';
+import { ErrorBoundary } from '../features/error-boundary/ErrorBoundary.tsx';
 
 interface HomeProps {
   characterData: CharacterData;
